1. add test coverage 2. migrate from nodeunit to jest
