+++ packages/@aws-cdk/aws-chatbot/package.json
@@ -48,7 +48,8 @@
     "compat": "cdk-compat"
   },
   "cdk-build": {
-    "cloudformation": "AWS::Chatbot"
+    "cloudformation": "AWS::Chatbot",
+    "jest": true
   },
   "keywords": [
     "aws",
@@ -68,9 +69,7 @@
     "cdk-build-tools": "0.0.0",
     "cdk-integ-tools": "0.0.0",
     "cfn2ts": "0.0.0",
-    "pkglint": "0.0.0",
-    "@types/nodeunit": "^0.0.31",
-    "nodeunit": "^0.11.3"
+    "pkglint": "0.0.0"
   },
   "dependencies": {
     "@aws-cdk/aws-iam": "0.0.0",
